move suites back to _*test
