+++ common/persistence/sql/sqlplugin/postgresql/driver/pgx.go
@@ -30,12 +30,6 @@ import (
 	"github.com/jmoiron/sqlx"
 )
 
-const (
-	// check http://www.postgresql.org/docs/9.3/static/errcodes-appendix.html
-	dupEntryCode    = "23505"
-	dupDatabaseCode = "42P04"
-)
-
 type PGXDriver struct{}
 
 func (p *PGXDriver) CreateConnection(dsn string) (*sqlx.DB, error) {
